Mental Health Chatbot/requirements.txt
Flask==2.2.3
requests==2.28.1
gunicorn==20.1.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
Werkzeug==2.2.3
python-dotenv==1.0.0